backend/requirements.txt
Django==3.2.10
django-filter==2.4.0
djangorestframework==3.12.4
djoser==2.2.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
